+++ Content.Shared/_CM14/Barricade/Components/BarbedWireComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Shared._CM14.Barricade.Components;
+
+[RegisterComponent]
+public sealed partial class BarbedWireComponent : Component
+{
+}
