--- Content.Shared/_CM14/Comtech/Barbed/Components/BarbedwireComponent.cs
@@ -1,6 +1,0 @@
-namespace Content.Shared._CM14.Comtech.Barbed.Components;
-
-[RegisterComponent]
-public sealed partial class BarbedwireComponent : Component
-{
-}
